+++ src/rdfTypes/GeoSparqlHelpers.cpp
@@ -16,10 +16,35 @@
 #include <string_view>
 #include <type_traits>
 
+#include "backports/algorithm.h"
 #include "global/Constants.h"
 #include "rdfTypes/GeoPoint.h"
 #include "rdfTypes/GeometryInfoHelpersImpl.h"
 #include "util/Exception.h"
+
+namespace {
+constexpr ctll::fixed_string de9imFilterRegex = "[0-2TtFf*]{9}";
+}  // namespace
+
+// ____________________________________________________________________________
+std::optional<De9imFilterString> parseDe9imFilterString(
+    std::string_view filter) {
+  if (!ctre::match<de9imFilterRegex>(filter)) {
+    return std::nullopt;
+  }
+  // The regex above already enforces that `filter` has exactly 9 characters.
+  AD_CORRECTNESS_CHECK(filter.size() == De9imFilterString{}.size());
+  De9imFilterString result{};
+  ql::ranges::copy(filter, result.begin());
+  return result;
+}
+
+// ____________________________________________________________________________
+bool de9imFilterCanMatchDisjoint(const De9imFilterString& filter) {
+  auto admitsF = [](char c) { return c == '*' || c == 'F' || c == 'f'; };
+  return admitsF(filter[0]) && admitsF(filter[1]) && admitsF(filter[3]) &&
+         admitsF(filter[4]);
+}
 
 namespace ad_utility {
