+++ test/QueryRewriteUtilTestHelpers.h
@@ -12,6 +12,7 @@
 #include "engine/sparqlExpressions/RelationalExpressions.h"
 #include "global/Constants.h"
 #include "parser/SpatialQuery.h"
+#include "rdfTypes/GeoSparqlHelpers.h"
 #include "rdfTypes/Iri.h"
 #include "util/GTestHelpers.h"
 #include "util/SourceLocation.h"
